requirements.txt
fastapi
uvicorn
python-multipart
PyMuPDF
spacy
sentence-transformers
openai
pydantic
scikit-learn
PyPDF2
python-dotenv